pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["Leo Kukharau <kuhorev.leonid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
django-crispy-forms = "^1.9.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
pylint-django = "^2.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
